+++ jsondataferret/tests/test_pythonapi_newevent.py
@@ -17,14 +17,18 @@ class NewEditCase(TestCase):
     def test_basic(self):
         """hello world"""
         data = NewEventData("animal", "lion", {"title": "Lion", "sound": "Roar!"})
+        assert {
+            "title": "Lion",
+            "sound": "Roar!",
+        } == data.get_new_data_when_event_applied_to_latest_record_cached_data()
         newEvent([data], None)
 
         edits = Edit.objects.all()
         assert 1 == len(edits)
         assert edits[0].data["sound"] == "Roar!"
 
-    def test_does_this_create_or_change_record_replace_no_change(self):
-        # Apply inital data - No record, so should be true
+    def test_replace_mode_with_no_change(self):
+        # Apply inital data
         data_initial = NewEventData(
             "animal",
             "lion",
@@ -32,9 +36,13 @@ class NewEditCase(TestCase):
             approved=True,
         )
         assert data_initial.does_this_create_or_change_record()
+        assert (
+            {"title": "Lion", "sound": "Roar!"}
+            == data_initial.get_new_data_when_event_applied_to_latest_record_cached_data()
+        )
         # Save it so we have record to compare more edits to
         newEvent([data_initial], None)
-        # Data with Mode Replace - same?
+        # Data with Mode Replace
         data_replace = NewEventData(
             "animal",
             "lion",
@@ -42,9 +50,13 @@ class NewEditCase(TestCase):
             mode=EVENT_MODE_REPLACE,
         )
         assert not data_replace.does_this_create_or_change_record()
+        assert (
+            {"title": "Lion", "sound": "Roar!"}
+            == data_replace.get_new_data_when_event_applied_to_latest_record_cached_data()
+        )
 
-    def test_does_this_create_or_change_record_merge_no_change(self):
-        # Apply inital data - No record, so should be true
+    def test_merge_mode_with_no_change(self):
+        # Apply inital data
         data_initial = NewEventData(
             "animal",
             "lion",
@@ -52,16 +64,24 @@ class NewEditCase(TestCase):
             approved=True,
         )
         assert data_initial.does_this_create_or_change_record()
+        assert (
+            {"title": "Lion", "sound": "Roar!"}
+            == data_initial.get_new_data_when_event_applied_to_latest_record_cached_data()
+        )
         # Save it so we have record to compare more edits to
         newEvent([data_initial], None)
-        # Data with Mode Merge - same?
+        # Data with Mode Merge
         data_merge = NewEventData(
             "animal", "lion", {"title": "Lion"}, mode=EVENT_MODE_MERGE
         )
         assert not data_merge.does_this_create_or_change_record()
+        assert (
+            {"title": "Lion", "sound": "Roar!"}
+            == data_merge.get_new_data_when_event_applied_to_latest_record_cached_data()
+        )
 
-    def test_does_this_create_or_change_record_replace_change(self):
-        # Apply inital data - No record, so should be true
+    def test_replace_mode_with_change(self):
+        # Apply inital data
         data_initial = NewEventData(
             "animal",
             "lion",
@@ -69,9 +89,13 @@ class NewEditCase(TestCase):
             approved=True,
         )
         assert data_initial.does_this_create_or_change_record()
+        assert (
+            {"title": "Lion", "sound": "Roar!"}
+            == data_initial.get_new_data_when_event_applied_to_latest_record_cached_data()
+        )
         # Save it so we have record to compare more edits to
         newEvent([data_initial], None)
-        # Data with Mode Replace - different?
+        # Data with Mode Replace
         data_replace = NewEventData(
             "animal",
             "lion",
@@ -79,9 +103,13 @@ class NewEditCase(TestCase):
             mode=EVENT_MODE_REPLACE,
         )
         assert data_replace.does_this_create_or_change_record()
+        assert (
+            {"title": "Big Cat", "sound": "Roar!"}
+            == data_replace.get_new_data_when_event_applied_to_latest_record_cached_data()
+        )
 
-    def test_does_this_create_or_change_record_merge_change(self):
-        # Apply inital data - No record, so should be true
+    def test_merge_mode_with_change(self):
+        # Apply inital data
         data_initial = NewEventData(
             "animal",
             "lion",
@@ -89,10 +117,18 @@ class NewEditCase(TestCase):
             approved=True,
         )
         assert data_initial.does_this_create_or_change_record()
+        assert (
+            {"title": "Lion", "sound": "Roar!"}
+            == data_initial.get_new_data_when_event_applied_to_latest_record_cached_data()
+        )
         # Save it so we have record to compare more edits to
         newEvent([data_initial], None)
-        # Data with Mode Merge - different?
+        # Data with Mode Merge
         data_merge = NewEventData(
             "animal", "lion", {"title": "Pretty Cat"}, mode=EVENT_MODE_MERGE
         )
         assert data_merge.does_this_create_or_change_record()
+        assert (
+            {"title": "Pretty Cat", "sound": "Roar!"}
+            == data_merge.get_new_data_when_event_applied_to_latest_record_cached_data()
+        )